feat : Hotel model API refactorig
멘토님 피드백 사항 반영. 각 기능마다 새 경로 지정. dataframe 함수 작성
파일 명 소문자 변경api로 확인할 경우
'uvicorn hotelAPI.main.hotelmain:app --reload' 경로 사용 바람.

footer : #6
