+++ HotelAPI/schemas/hotel_schema.py
